+++ services/ui-src/src/measures/2022/shared/globalValidations/index.ts
@@ -9,7 +9,7 @@ export * from "shared/globalValidations/validateBothDatesInRange";
 export * from "./validateDualPopInformation";
 export * from "shared/globalValidations/validateEqualCategoryDenominators";
 export * from "shared/globalValidations/validateEqualQualifierDenominators";
-export * from "./validateRateNotZero";
+export * from "shared/globalValidations/validateRateNotZero";
 export * from "./validateRateZero";
 export * from "shared/globalValidations/validateNumeratorsLessThanDenominators";
 export * from "shared/globalValidations/validateOneCatRateHigherThanOtherCat";
