+++ services/ui-src/src/measures/2022/shared/util/validationsMock.tsx
@@ -5,7 +5,7 @@ import * as validateBothDatesInRange from "shared/globalValidations/validateBoth
 import * as validateDualPopInformation from "measures/2022/shared/globalValidations/validateDualPopInformation";
 import * as validateEqualCategoryDenominators from "shared/globalValidations/validateEqualCategoryDenominators";
 import * as validateEqualQualifierDenominators from "shared/globalValidations/validateEqualQualifierDenominators";
-import * as validateRateNotZero from "measures/2022/shared/globalValidations/validateRateNotZero";
+import * as validateRateNotZero from "shared/globalValidations/validateRateNotZero";
 import * as validateRateZero from "measures/2022/shared/globalValidations/validateRateZero";
 import * as validateNumeratorsLessThanDenominators from "shared/globalValidations/validateNumeratorsLessThanDenominators";
 import * as validateOneCatRateHigherThanOtherCat from "shared/globalValidations/validateOneCatRateHigherThanOtherCat";
