+++ services/ui-src/src/measures/2023/shared/globalValidations/index.ts
@@ -14,7 +14,7 @@ export * from "./validateDualPopInformation";
 export * from "shared/globalValidations/validateEqualCategoryDenominators";
 export * from "shared/globalValidations/validateEqualQualifierDenominators";
 export * from "shared/globalValidations/validateFfsRadioButtonCompletion";
-export * from "./validateRateNotZero";
+export * from "shared/globalValidations/validateRateNotZero";
 export * from "./validateRateZero";
 export * from "shared/globalValidations/validateNumeratorsLessThanDenominators";
 export * from "shared/globalValidations/validateOneCatRateHigherThanOtherCat";
