+++ services/ui-src/src/measures/2024/shared/util/validationsMock.tsx
@@ -9,7 +9,7 @@ import * as validateDualPopInformation from "measures/2024/shared/globalValidati
 import * as validateEqualCategoryDenominators from "shared/globalValidations/validateEqualCategoryDenominators";
 import * as validateEqualQualifierDenominators from "shared/globalValidations/validateEqualQualifierDenominators";
 import * as validateFfsRadioButtonCompletion from "shared/globalValidations/validateFfsRadioButtonCompletion";
-import * as validateRateNotZero from "measures/2024/shared/globalValidations/validateRateNotZero";
+import * as validateRateNotZero from "shared/globalValidations/validateRateNotZero";
 import * as validateRateZero from "measures/2024/shared/globalValidations/validateRateZero";
 import * as validateNumeratorsLessThanDenominators from "shared/globalValidations/validateNumeratorsLessThanDenominators";
 import * as validateOneCatRateHigherThanOtherCat from "shared/globalValidations/validateOneCatRateHigherThanOtherCat";
