+++ services/ui-src/src/shared/globalValidations/validateRateNotZero/index.test.ts
